requirement.txt
annotated-types==0.7.0    
anyio==4.9.0    
certifi==2025.7.14
charset-normalizer==3.4.2    
colorama==0.4.6    
distro==1.9.0    
exceptiongroup==1.3.0    
h11==0.16.0   
httpcore==1.0.9    
httpx==0.28.1   
idna==3.10     
jiter==0.10.0   
openai==1.97.0   
pip==21.3.1   
pydantic==2.11.7   
pydantic_core==2.33.2   
python-dotenv==1.1.1    
requests==2.32.4   
setuptools==60.2.0   
sniffio==1.3.1    
tqdm==4.67.1   
typing_extensions==4.14.1   
typing-inspection==0.4.1    
urllib3==2.5.0    
wheel==0.37.1